chore: 쿼리 index 추가
